requirements.txt
noise>=1.2.1
pytest
matplotlib
numpy
PyYAML
